feat: add virtual scrolling. add chunking of data into rows.
got the data chunked so each row can show multiple cards. The issue is, that's currently static. I need to be able to find the width of the virtual viewport and then do some quick math to figure out how many cards can go in a row based on that.